chore: move File class and add name property
