+++ src/modules/shared/util/File/index.ts
@@ -1,0 +1,1 @@
+export * from './File';
